Memory: Split into Engine and GameDLL files
